Backend/requirements.txt
Django==3.2.9
mysqlclient==2.1.1
djangorestframework==3.13.1
geopy==2.2.0
django-cors-headers==3.13.0
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-filter==2.4.0
django-storages==1.12.3
django-environ==0.7.0